Give option to set validator URL (#27)
* Add validation URL to Swagger UI configuration

A new option 'validation_url' has been added in the 'swagger-ui.php' configuration file and also in the 'index.blade.php'. This will allow users to specify a URL for validation of the Swagger UI. If no URL is provided, then the validation will not run. This was necessary to offer the possibility of using an external validator if needed.

* Add validation URL to Swagger UI configuration

A new option 'validation_url' has been added in the 'swagger-ui.php' configuration file and also in the 'index.blade.php'. This will allow users to specify a URL for validation of the Swagger UI. If no URL is provided, then the validation will not run. This was necessary to offer the possibility of using an external validator if needed.

* Check for key existence in $data array

* Revert key existence checking

* Provide validator URL

---------

Co-authored-by: knorbert <kocsisnorbert@sbagroupzrt.onmicrosoft.com>